RQ2/mstracker/requirements.txt
cycler==0.10.0
gitdb2==3.0.2
GitPython==3.0.8
kiwisolver==1.1.0
lxml==4.5.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
peewee==3.13.1
psycopg2-binary==2.8.5
pyparsing==2.4.6
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
six==1.14.0
smmap2==2.0.5
